WIP: Started adding tests for the dataporten security stuff. OAuth2 process flow is more or less complete. Missing "touch" for session activity.

+++ musit-service/src/main/scala/no/uio/musit/security/oauth2/OAuth2Info.scala
@@ -26,7 +26,7 @@ import play.api.libs.functional.syntax._
 case class OAuth2Info(
   accessToken: BearerToken,
   tokenType: Option[String] = None,
-  expiresIn: Option[Int] = None,
+  expiresIn: Option[Long] = None,
   refreshToken: Option[String] = None,
   params: Option[Map[String, String]] = None
 )
@@ -36,7 +36,7 @@ object OAuth2Info extends OAuth2Constants {
   implicit val reads: Reads[OAuth2Info] = (
     (__ \ AccessToken).read[BearerToken] and
     (__ \ TokenType).readNullable[String] and
-    (__ \ ExpiresIn).readNullable[Int] and
+    (__ \ ExpiresIn).readNullable[Long] and
     (__ \ RefreshToken).readNullable[String]
   )((accessToken, tokenType, expiresIn, refreshToken) =>
       OAuth2Info(accessToken, tokenType, expiresIn, refreshToken))
